% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsSum.R
\name{sumbms}
\alias{sumbms}
\title{Bayesian hierarchical models for the combining the summay measures
\code{sumbms} returns the posterior distributions of parameters fitted with data \code{data}}
\usage{
sumbms(data, dim = 1, with_cov = FALSE, bi_option = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{dim}{A number 1 or 2.}

\item{with_cov}{logical, TRUE or FALSE}

\item{bi_option}{logical, TRUE or FALSE}
}
\value{
The posterior distributions of parameters fitted with data \code{data} with univariate or bivariate model.
}
\description{
Bayesian hierarchical models for the combining the summay measures
\code{sumbms} returns the posterior distributions of parameters fitted with data \code{data}
}
\examples{
add(data, 1)
add(data, 2)
}
